Move open/close modal to modal.js
Move state from utils to constants
