+++ src/js/utils/helpers.js
@@ -81,25 +81,6 @@ function sortCustomersByName(customers, currentSortingState) {
   }
 }
 
-//* Open and close modal functions
-function openModal(modal) {
-  const modalElement = modal;
-  const modalOverlay = document.querySelector('.modal-overlay');
-  modalElement.classList.add('open'); // Show the modal
-  modalOverlay.classList.add('open'); // Show the overlay
-}
-
-function closeModal(modal) {
-  const modalElement = modal;
-  const modalOverlay = document.querySelector('.modal-overlay');
-  modalElement.classList.remove('open'); // Hide the modal
-  while (modalElement.firstChild) {
-    modalElement.removeChild(modalElement.firstChild);
-  }
-
-  modalOverlay.classList.remove('open'); // Hide the overlay
-}
-
 // Add/Edit form validation
 function checkFormValidity() {
   const confirmButton = document.querySelector('.confirm-button');
@@ -121,7 +102,5 @@ export {
   combineAndRemoveDuplicates,
   getActionMenuPosition,
   sortCustomersByName,
-  openModal,
-  closeModal,
   checkFormValidity,
 };
